# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Shorty URL shortener",
        "title": "Shorty API",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/analytics": {
            "get": {
                "description": "Returns all click analytics grouped by short URLs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get all click analytics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/analytics.Analysis"
                            }
                        }
                    },
                    "404": {
                        "description": "No clicks found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch analytics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Records a click for a short URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "CreateClick a new click record",
                "parameters": [
                    {
                        "description": "Click information",
                        "name": "click",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/analytics.ClickEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/analytics.ClickModel"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create click",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/clicks": {
            "get": {
                "description": "Returns all individual click records (not grouped)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get all click records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/analytics.ClickModel"
                            }
                        }
                    },
                    "404": {
                        "description": "No clicks found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch clicks",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/clicks/{id}": {
            "get": {
                "description": "Returns a single click record based on its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get a click record by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Click ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analytics.ClickModel"
                        }
                    },
                    "400": {
                        "description": "Invalid ID parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Click not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch click",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/{shortUrl}": {
            "get": {
                "description": "Returns click analytics for the given short URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get analytics for a specific short URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL identifier",
                        "name": "shortUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analytics.Analysis"
                        }
                    },
                    "400": {
                        "description": "Missing shortUrl parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No clicks found for this short URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch clicks",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "analytics.Analysis": {
            "type": "object",
            "properties": {
                "originalUrl": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "usageDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/analytics.Usage"
                    }
                }
            }
        },
        "analytics.ClickEvent": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "short_id": {
                    "type": "integer"
                },
                "time_stamp": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "analytics.ClickModel": {
            "type": "object",
            "required": [
                "ip_address",
                "short_id",
                "user_agent"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_address": {
                    "type": "string"
                },
                "short": {
                    "$ref": "#/definitions/shortener.ShortModel"
                },
                "short_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "analytics.Usage": {
            "type": "object",
            "properties": {
                "clickTimes": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "userAgents": {
                    "type": "string"
                }
            }
        },
...
        "shortener.ShortModel": {
            "type": "object",
            "required": [
                "originalUrl",
                "shortUrl",
                "userID"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "originalUrl": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
    }
}